+++ src/lib.rs
@@ -71,15 +71,13 @@
 extern crate proc_macro;
 
 use proc_macro::TokenStream;
-use proc_macro2::TokenStream as TokenStream2;
-use quote::ToTokens;
-use syn::token;
+use quote::{quote, ToTokens};
 use syn_mid::ItemFn;
 
 /// An attribute for easy generation of a const function with conditional compilations.
 #[proc_macro_attribute]
 pub fn const_fn(args: TokenStream, function: TokenStream) -> TokenStream {
-    let args = TokenStream2::from(args);
+    let args = proc_macro2::TokenStream::from(args);
 
     if args.is_empty() {
         return syn::Error::new_spanned(args, "`const_fn` requires an argument")
@@ -87,20 +85,23 @@ pub fn const_fn(args: TokenStream, function: TokenStream) -> TokenStream {
             .into();
     }
 
-    let mut function: ItemFn = syn::parse_macro_input!(function);
+    let mut item: ItemFn = syn::parse_macro_input!(function);
 
-    let mut const_function = function.clone();
-
-    if function.constness.is_some() {
-        function.constness = None;
-    } else {
-        const_function.constness = Some(token::Const::default());
+    if item.constness.is_none() {
+        return syn::Error::new_spanned(
+            item.fn_token,
+            "#[const_fn] attribute may only be used on const functions",
+        )
+        .to_compile_error()
+        .into();
     }
 
-    function.attrs.push(syn::parse_quote!(#[cfg(not(#args))]));
-    const_function.attrs.push(syn::parse_quote!(#[cfg(#args)]));
+    let mut token = quote!(#[cfg(#args)]);
+    token.extend(item.to_token_stream());
 
-    let mut function = function.into_token_stream();
-    function.extend(const_function.into_token_stream());
-    TokenStream::from(function)
+    item.attrs.push(syn::parse_quote!(#[cfg(not(#args))]));
+    item.constness = None;
+    token.extend(item.into_token_stream());
+
+    token.into()
 }