+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* `#[const_fn]` attribute may only be used on const functions.
+
 # 0.2.1 - 2019-08-15
 
 * Updated `proc-macro2`, `syn`, and `quote` to 1.0.